# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Maicare server API documentation.",
        "title": "Maicare API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "your-email@domain.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/clients/{id}/progress_reports": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress_reports"
                ],
                "summary": "List all progress reports for a client",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Client ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-pagination_Response-array_api_ListProgressReportsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress_reports"
                ],
                "summary": "Create a new progress report for a client",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Client ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Progress Report Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateProgressReportRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.Response-api_CreateProgressReportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/clients/{id}/progress_reports/{report_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress_reports"
                ],
                "summary": "Retrieve a progress report for a client",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Client ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Progress Report ID",
                        "name": "report_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-api_GetProgressReportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "progress_reports"
                ],
                "summary": "Update a progress report for a client",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Client ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Progress Report ID",
                        "name": "report_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Progress Report Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateProgressReportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-api_UpdateProgressReportResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.CreateProgressReportRequest": {
            "type": "object",
            "required": [
                "emotional_state",
                "report_text",
                "type"
            ],
            "properties": {
                "date": {
                    "type": "string"
                },
                "emotional_state": {
                    "type": "string",
                    "enum": [
                        "normal",
                        "excited",
                        "happy",
                        "sad",
                        "angry",
                        "anxious",
                        "depressed"
                    ]
                },
                "employee_id": {
                    "type": "integer"
                },
                "report_text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "morning_report",
                        "evening_report",
                        "night_report",
                        "shift_report",
                        "one_to_one_report",
                        "process_report",
                        "contact_journal",
                        "other"
                    ]
                }
            }
        },
        "api.CreateProgressReportResponse": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "emotional_state": {
                    "type": "string"
                },
                "employee_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "report_text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "api.GetProgressReportResponse": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "emotional_state": {
                    "type": "string"
                },
                "employee_first_name": {
                    "type": "string"
                },
                "employee_id": {
                    "type": "integer"
                },
                "employee_last_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "report_text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "api.ListProgressReportsResponse": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "emotional_state": {
                    "type": "string"
                },
                "employee_first_name": {
                    "type": "string"
                },
                "employee_id": {
                    "type": "integer"
                },
                "employee_last_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "report_text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "api.Response-any": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.Response-api_CreateProgressReportResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/api.CreateProgressReportResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.Response-api_GetProgressReportResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/api.GetProgressReportResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.Response-api_UpdateProgressReportResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/api.UpdateProgressReportResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.Response-pagination_Response-array_api_ListProgressReportsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/pagination.Response-array_api_ListProgressReportsResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.UpdateProgressReportRequest": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "emotional_state": {
                    "type": "string"
                },
                "employee_id": {
                    "type": "integer"
                },
                "report_text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "api.UpdateProgressReportResponse": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "emotional_state": {
                    "type": "string"
                },
                "employee_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "report_text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "pagination.Response-array_api_ListProgressReportsResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "next": {
                    "type": "string"
                },
                "page_size": {
                    "type": "integer"
                },
                "previous": {
                    "type": "string"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/api.ListProgressReportsResponse"
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Add 'Bearer ' prefix before your JWT token for authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}